[FIX] 기타 버그 및 코드 수정

+++ Source/GA6thFinal_Framework/GraphicsEngine/ParticleManager.cpp
@@ -797,7 +797,7 @@ void ParticleManager::AwakeParticles(float deltaTime, ParticleUpdateResource* sc
         }
     }
     // 최종 파티클 수 업데이트
-    scene->_totalCount = scene->_totalParticles.size();
+    scene->_totalCount = (UINT)scene->_totalParticles.size();
     scene->_ribbonTotalCount = (UINT)(scene->_ribbonTotalParticles.size());
 }
